polymarket/requirements.txt
pyyaml==6.0.2
pydantic==2.9.2
httpx==0.27.2
websockets==13.1
tenacity==9.0.0